requirements.txt
asgiref==3.4.0
dj-database-url==0.5.0
Django==3.2.4
django-heroku==0.3.1
gunicorn==20.1.0
install==1.3.4
psycopg2==2.9.1
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
django-cors-headers
django-debug-toolbar
python-dotenv 
django-split-settings
whitenoise
djangorestframework-api-key
django-filter
djangorestframework-word-filter
djangorestframework-jwt
pyyaml 
django-rest-swagger
uritemplate